+++ package.json
@@ -30,6 +30,7 @@
     "@react-native-picker/picker": "2.6.1",
     "@react-navigation/drawer": "^6.1.8",
     "@react-navigation/native": "^6.0.6",
+    "base-64": "^1.0.0",
     "moti": "^0.16.1",
     "native-base": "^3.2.1",
     "normalize-css-color": "^1.0.2",
@@ -49,15 +50,15 @@
     "styled-system": "^5.1.5"
   },
   "devDependencies": {
+    "@babel/core": "^7.12.9",
+    "@expo/metro-runtime": "~3.1.3",
+    "@types/react": "~18.2.45",
+    "@types/react-native": "~0.64.12",
+    "@types/shortid": "^0.0.29",
     "expo": "^50.0.6",
     "expo-linking": "~6.2.2",
     "expo-modules-core": "~1.11.8",
     "expo-status-bar": "~1.11.1",
-    "@expo/metro-runtime": "~3.1.3",
-    "@babel/core": "^7.12.9",
-    "@types/react": "~18.2.45",
-    "@types/react-native": "~0.64.12",
-    "@types/shortid": "^0.0.29",
     "prettier": "^2.4.1",
     "typescript": "^5.3.0"
   },
